+++ .gitignore
@@ -1,1 +1,2 @@
 log/log.txt
+include/config/DbManager.php
