+++ include/model/encode.php
@@ -1,0 +1,5 @@
+<?php
+//エスケープ
+function e(string $str, string $charset = 'UTF-8'):string {
+    return htmlspecialchars($str, ENT_QUOTES | ENT_HTML5, $charset);
+}
